refactor: change module name to azion ai
