+++ src/assets/svg/ai-ask-azion.vue
@@ -1,0 +1,22 @@
+<template>
+  <svg
+    width="14"
+    height="15"
+    viewBox="0 0 14 15"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M13.3795 2.09575C13.2823 1.89363 13.0779 1.76508 12.8537 1.76508C12.6294 1.76508 12.425 1.89363 12.3279 2.09575C12.2883 2.17815 12.2529 2.24127 12.2191 2.28717C12.1975 2.31664 12.1841 2.32854 12.1808 2.33137C11.9915 2.43286 11.8731 2.63034 11.8731 2.84546C11.8731 3.06058 11.9915 3.25806 12.1808 3.35955C12.1841 3.36238 12.1975 3.37427 12.2191 3.40375C12.2529 3.44964 12.2883 3.51276 12.3279 3.59517C12.425 3.79729 12.6294 3.92584 12.8537 3.92584C13.0779 3.92584 13.2823 3.79729 13.3795 3.59517C13.4191 3.51276 13.4545 3.44964 13.4882 3.40375C13.5099 3.37427 13.5233 3.36238 13.5266 3.35955C13.7159 3.25806 13.8342 3.06058 13.8342 2.84546C13.8342 2.63034 13.7159 2.43286 13.5266 2.33137C13.5233 2.32854 13.5099 2.31664 13.4882 2.28717C13.4545 2.24127 13.4191 2.17815 13.3795 2.09575ZM5.91955 3.37823C3.82075 3.37823 2.11956 5.07923 2.11956 7.17725C2.11956 7.5814 2.18248 7.96973 2.29864 8.33362C2.33651 8.45225 2.33542 8.57988 2.29553 8.69784L1.46845 11.1441L3.95123 10.5866C4.07655 10.5584 4.20765 10.5724 4.3242 10.6264C4.80857 10.8508 5.3486 10.9763 5.91955 10.9763C6.11282 10.9763 6.30236 10.9619 6.48724 10.9342C6.80586 10.8866 7.10279 11.1062 7.15044 11.4249C7.19809 11.7435 6.97843 12.0404 6.65981 12.0881C6.41799 12.1242 6.17077 12.1429 5.91955 12.1429C5.24784 12.1429 4.60604 12.0093 4.0203 11.7668L0.711144 12.5098C0.50543 12.556 0.2908 12.4876 0.149757 12.3309C0.0087141 12.1742 -0.0367878 11.9536 0.0307421 11.7538L1.13107 8.49948C1.01484 8.07791 0.952905 7.63439 0.952905 7.17725C0.952905 4.43466 3.17667 2.21157 5.91955 2.21157C6.44253 2.21157 6.9477 2.29259 7.42257 2.44318C7.72966 2.54056 7.89966 2.86845 7.80228 3.17555C7.7049 3.48264 7.377 3.65264 7.06991 3.55526C6.70776 3.44041 6.32147 3.37823 5.91955 3.37823ZM9.67907 3.3368C9.90628 3.3368 10.1128 3.46874 10.2083 3.6749C10.3558 3.99318 10.4615 4.40435 10.5538 4.76368L10.5538 4.76378L10.5539 4.76395C10.5815 4.87168 10.608 4.97474 10.6341 5.06925C10.763 5.53691 10.8986 5.88918 11.0985 6.11871C11.2722 6.31806 11.6348 6.48703 12.1763 6.66091C12.3027 6.70151 12.431 6.74022 12.562 6.77959L12.5878 6.78734C12.7089 6.82371 12.833 6.86098 12.9524 6.89893C13.202 6.97821 13.4694 7.07195 13.6946 7.194C13.8828 7.29597 14 7.49282 14 7.70686C14 7.9209 13.8828 8.11775 13.6946 8.21972C13.4707 8.34108 13.2043 8.43557 12.9555 8.51615C12.8214 8.55956 12.7073 8.59457 12.5954 8.62892L12.5954 8.62892L12.5953 8.62894C12.4639 8.66928 12.3354 8.70871 12.1811 8.75974C11.6417 8.93814 11.2776 9.11253 11.1023 9.31786C10.8999 9.55494 10.7636 9.91575 10.6345 10.3889C10.6087 10.4835 10.5825 10.5864 10.5552 10.6937C10.4625 11.0576 10.3567 11.473 10.2083 11.7933C10.1128 11.9995 9.90628 12.1314 9.67907 12.1314C9.45185 12.1314 9.24531 11.9995 9.14979 11.7933C9.00139 11.473 8.89561 11.0576 8.80295 10.6937L8.80287 10.6934C8.77557 10.5862 8.7494 10.4834 8.72362 10.3889C8.59457 9.91575 8.45827 9.55494 8.25583 9.31786C8.0805 9.11253 7.71644 8.93814 7.17705 8.75974C7.02274 8.70871 6.89427 8.66928 6.76279 8.62894C6.65085 8.59459 6.53673 8.55957 6.40264 8.51615C6.15381 8.43557 5.88747 8.34108 5.66353 8.21972C5.47535 8.11775 5.35813 7.9209 5.35813 7.70686C5.35813 7.49282 5.47535 7.29597 5.66353 7.194C5.88875 7.07195 6.15617 6.97821 6.40569 6.89893C6.52514 6.86098 6.64922 6.82371 6.77031 6.78734L6.79612 6.77959C6.92713 6.74022 7.05542 6.70151 7.18188 6.66091C7.72338 6.48703 8.08594 6.31806 8.2596 6.11871C8.45954 5.88918 8.59511 5.53691 8.72405 5.06925C8.75012 4.9747 8.77661 4.87158 8.80431 4.76379L8.80434 4.76369C8.89667 4.40435 9.00232 3.99318 9.14979 3.6749C9.24531 3.46874 9.45185 3.3368 9.67907 3.3368ZM9.67907 5.92024C9.55637 6.25536 9.38889 6.59849 9.13929 6.88502C8.77242 7.30616 8.20432 7.547 7.71968 7.71196C8.2067 7.88249 8.77631 8.13078 9.14305 8.56028C9.39049 8.85006 9.5569 9.19578 9.67907 9.53343C9.80123 9.19578 9.96764 8.85006 10.2151 8.56028C10.5818 8.13078 11.1514 7.88249 11.6385 7.71196C11.1538 7.547 10.5857 7.30616 10.2188 6.88502C9.96924 6.59849 9.80176 6.25536 9.67907 5.92024Z"
+      fill="#F4F4F4"
+    />
+  </svg>
+</template>
+
+<script setup>
+  defineOptions({
+    name: 'ai-azion-avatar-svg'
+  })
+</script>
