--- src/modules/ask-azion/services/requestInterceptorService.js
@@ -1,5 +1,0 @@
-export const requestInterceptorService = (requestDetails, { sessionId, url }) => {
-  requestDetails.body['session_id'] = sessionId
-  requestDetails.body['url'] = url
-  return requestDetails
-}
