+++ src/templates/navbar-block/ai-chat-button.vue
@@ -13,20 +13,20 @@
     icon="pi pi-sparkles"
     size="small"
     class="text-white border-header"
-    v-tooltip.bottom="{ value: 'Ask Azion', showDelay: 200 }"
+    v-tooltip.bottom="{ value: 'Azion AI', showDelay: 200 }"
   />
 </template>
 
 <script setup>
   import { computed, inject } from 'vue'
-  import { useAskAzionAiChatStore } from '@/stores/ask-azion-ai-chat'
+  import { useAzionAiChatStore } from '@/stores/azion-ai-chat-store'
   import { useHelpCenterStore } from '@/stores/help-center'
 
   import PrimeButton from 'primevue/button'
 
   defineOptions({ name: 'ai-chat-button' })
 
-  const askAzionAiChatStore = useAskAzionAiChatStore()
+  const askAzionAiChatStore = useAzionAiChatStore()
   const helpCenterStore = useHelpCenterStore()
   const currentWidth = inject('currentWidth')
   const SCREEN_BREAKPOINT_MD = 768
@@ -38,7 +38,7 @@
 
   const currentLabel = computed(() => {
     if (currentWidth.value > SCREEN_BREAKPOINT_MD) {
-      return 'Ask Azion'
+      return 'Azion AI'
     }
     return ''
   })
@@ -47,3 +47,4 @@
     return askAzionAiChatStore.isOpen ? 'active-helper' : ''
   })
 </script>
+@/stores/azion-ai-chat-store
