+++ src/templates/navbar-block/help.vue
@@ -20,13 +20,13 @@
 <script setup>
   import { computed, inject } from 'vue'
   import { useHelpCenterStore } from '@/stores/help-center'
-  import { useAskAzionAiChatStore } from '@/stores/ask-azion-ai-chat'
+  import { useAzionAiChatStore } from '@/stores/azion-ai-chat-store'
   import PrimeButton from 'primevue/button'
 
   defineOptions({ name: 'navbar-help-block' })
 
   const helpCenterStore = useHelpCenterStore()
-  const askAzionAiChatStore = useAskAzionAiChatStore()
+  const askAzionAiChatStore = useAzionAiChatStore()
   const currentWidth = inject('currentWidth')
   const SCREEN_BREAKPOINT_MD = 768
 
@@ -46,3 +46,4 @@
     return helpCenterStore.isOpen
   })
 </script>
+@/stores/azion-ai-chat-store
